+++ realm/realm-library/src/androidTest/java/io/realm/RealmTests.java
@@ -4322,6 +4322,7 @@ public class RealmTests {
                 fail();
             }
         });
+        looperThread.testComplete();
     }
 
     @Test
@@ -4329,6 +4330,7 @@ public class RealmTests {
     public void getInstanceAsync_nullCallbackShouldThrow() {
         thrown.expect(IllegalArgumentException.class);
         Realm.getInstanceAsync(realmConfig, null);
+        looperThread.testComplete();
     }
 
     // Verify that the logic for waiting for the users file dir to be come available isn't totally broken