+++ realm/realm-library/src/androidTestObjectServer/java/io/realm/SyncManagerTests.java
@@ -95,6 +95,8 @@ public class SyncManagerTests {
             userStore.remove(syncUser.getIdentity(), syncUser.getAuthenticationUrl().toString());
         }
         SyncManager.reset();
+        BaseRealm.applicationContext = null; // Required for Realm.init() to work
+        Realm.init(InstrumentationRegistry.getTargetContext());
     }
 
     @Test