+++ realm/realm-library/src/main/cpp/util.cpp
@@ -30,8 +30,13 @@
 #include "results.hpp"
 #include "list.hpp"
 #include "java_exception_def.hpp"
+#if REALM_ENABLE_SYNC
+#include "sync/partial_sync.hpp"
+#endif
 
 #include "jni_util/java_exception_thrower.hpp"
+
+
 
 using namespace std;
 using namespace realm;
@@ -120,6 +125,14 @@ void ConvertException(JNIEnv* env, const char* file, int line)
         }
         ThrowException(env, kind, e.what());
     }
+#if REALM_ENABLE_SYNC
+    catch (partial_sync::InvalidRealmStateException& e) {
+        ThrowException(env, IllegalState, e.what());
+    }
+    catch (partial_sync::ExistingSubscriptionException& e) {
+        ThrowException(env, IllegalArgument, e.what());
+    }
+#endif
     catch (std::logic_error e) {
         ThrowException(env, IllegalState, e.what());
     }