+++ realm/realm-library/src/main/java/io/realm/Realm.java
@@ -72,6 +72,7 @@ import io.realm.internal.Util;
 import io.realm.internal.annotations.ObjectServer;
 import io.realm.internal.async.RealmAsyncTaskImpl;
 import io.realm.log.RealmLog;
+import io.realm.sync.Subscription;
 import io.realm.sync.permissions.ClassPermissions;
 import io.realm.sync.permissions.ClassPrivileges;
 import io.realm.sync.permissions.RealmPermissions;
@@ -1864,6 +1865,47 @@ public class Realm extends BaseRealm {
         return where(ClassPermissions.class)
                 .equalTo("name", configuration.getSchemaMediator().getSimpleClassName(clazz))
                 .findFirst();
+    }
+
+    /**
+     * Returns a list of all known subscriptions, regardless of their status.
+     *
+     * @return a list of all known subscriptions.
+     */
+    @Beta
+    @ObjectServer
+    public RealmResults<Subscription> getSubscriptions() {
+        return where(Subscription.class).findAll();
+    }
+
+    /**
+     * Returns a list of all subscriptions that match a given pattern. {@code *} can be used to
+     * indicate any number of unknown characters and {@code ?} represents a single unknown character.
+     *
+     * @param pattern which subscriptions to find.
+     * @return list of subscriptions that match the pattern.
+     * @throws IllegalArgumentException if an empty or {@code null} pattern is provided.
+     */
+    @Beta
+    @ObjectServer
+    public RealmResults<Subscription> getSubscriptions(String pattern) {
+        if (Util.isEmptyString(pattern)) {
+            throw new IllegalArgumentException("Non-empty 'pattern' required");
+        }
+        return where(Subscription.class).like("name", pattern).findAll();
+    }
+
+    /**
+     * Returns the first subscription that matches the given name.
+     *
+     * @param name the name of the subscription to find.
+     * @return returns the subscription that matches the name or {@code null} if no subscription matches the name.
+     */
+    @Beta
+    @ObjectServer
+    @Nullable
+    public Subscription getSubscription(String name) {
+        return where(Subscription.class).equalTo("name", name).findFirst();
     }
 
     Table getTable(Class<? extends RealmModel> clazz) {